% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{barcode_based_deconvolve}
\alias{barcode_based_deconvolve}
\title{Identify high-resolution barcodes to discard based on low-resolution barcodes}
\usage{
barcode_based_deconvolve(barcode_list, low_res, high_res)
}
\arguments{
\item{barcode_list}{A character vector of low-resolution barcodes}

\item{low_res}{Integer value indicating low-resolution bin size (e.g., 64)}

\item{high_res}{Integer value indicating high-resolution bin size (e.g., 8)}
}
\value{
A character vector of high-resolution barcodes to discard
}
\description{
Identify high-resolution barcodes to discard based on low-resolution barcodes
}
