% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{barcodes_to_pseudo_coordinates}
\alias{barcodes_to_pseudo_coordinates}
\title{Split barcodes into pseudo-coordinates  (e.g., "s_008um_00792_00584-1" -> "792, 584")}
\usage{
barcodes_to_pseudo_coordinates(barcode_list)
}
\arguments{
\item{barcode_list}{A character vector of Visium barcodes (e.g., "s_008um_00792_00584-1")}
}
\value{
A data.frame containing the original barcode and extracted row and col values
}
\description{
Split barcodes into pseudo-coordinates  (e.g., "s_008um_00792_00584-1" -> "792, 584")
}
