% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{estimate_deconvolved_scores}
\alias{estimate_deconvolved_scores}
\title{Estimate agreement scores between barcode-based and coordinate-based methods}
\usage{
estimate_deconvolved_scores(
  df,
  parquet_low_res,
  parquet_high_res,
  json_low_res,
  low_res,
  high_res
)
}
\arguments{
\item{df}{A data.frame that must contain a column named 'barcodes' (result of barcodes_to_pseudo_coordinates)}

\item{parquet_low_res}{Path to the low-resolution .parquet file}

\item{parquet_high_res}{Path to the high-resolution .parquet file}

\item{json_low_res}{Path to the low-resolution scale factors .json file}

\item{low_res}{Integer value indicating low-resolution bin size}

\item{high_res}{Integer value indicating high-resolution bin size}
}
\value{
Returns the input data.frame with an added 'scores' column;
  also prints a density plot of scores and scatter plots for lowest scores
}
\description{
Estimate agreement scores between barcode-based and coordinate-based methods
}
