% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_spatial_qc}
\alias{plot_spatial_qc}
\title{Plot spatial QC overlaying spots on the background image}
\usage{
plot_spatial_qc(
  spe,
  sample_id = "sample_id",
  sample = unique(colData(spe)[[sample_id]])[1],
  metric = "detected",
  outliers = NULL,
  point_size = 2,
  colors = c("white", "black"),
  stroke = 1
)
}
\arguments{
\item{spe}{SpatialExperiment object or similar containing spatial coordinates and images}

\item{sample_id}{Character string specifying the column in colData that contains sample IDs}

\item{sample}{Character string specifying which sample to plot (defaults to the first unique value)}

\item{metric}{Character string specifying the column in colData to color by (e.g., "detected")}

\item{outliers}{NULL or character string specifying the column in colData that marks outliers/highlight groups. Can be boolean or factor.}

\item{point_size}{Numeric value for spot size}

\item{colors}{A character vector of length 2 (gradient) or more for fill colors}

\item{stroke}{Numeric value for border size of highlighted spots}
}
\value{
A ggplot object
}
\description{
Plot spatial QC overlaying spots on the background image
}
